Rename daemon's modules and structs for clarity
