+++ TODO
@@ -1,3 +1,4 @@
 Fully automated, complete testing, for every option
 
 Revise daemon code
+	- reimplement syncing animations
